+++ bt/composites/bt_random_sequence.cpp
@@ -13,7 +13,7 @@ void BTRandomSequence::_enter() {
 	indicies.shuffle();
 }
 
-int BTRandomSequence::_tick(float p_delta) {
+int BTRandomSequence::_tick(double p_delta) {
 	int status = SUCCESS;
 	for (int i = last_running_idx; i < get_child_count(); i++) {
 		status = get_child(indicies[i])->execute(p_delta);
