+++ bt/composites/bt_selector.cpp
@@ -6,7 +6,7 @@ void BTSelector::_enter() {
 	last_running_idx = 0;
 }
 
-int BTSelector::_tick(float p_delta) {
+int BTSelector::_tick(double p_delta) {
 	int status = FAILURE;
 	for (int i = last_running_idx; i < get_child_count(); i++) {
 		status = get_child(i)->execute(p_delta);
