+++ bt/decorators/bt_always_succeed.cpp
@@ -2,7 +2,7 @@
 
 #include "bt_always_succeed.h"
 
-int BTAlwaysSucceed::_tick(float p_delta) {
+int BTAlwaysSucceed::_tick(double p_delta) {
 	if (get_child_count() > 0 && get_child(0)->execute(p_delta) == RUNNING) {
 		return RUNNING;
 	}
