+++ bt/decorators/bt_repeat.cpp
@@ -12,7 +12,7 @@ void BTRepeat::_enter() {
 	cur_iteration = 1;
 }
 
-int BTRepeat::_tick(float p_delta) {
+int BTRepeat::_tick(double p_delta) {
 	ERR_FAIL_COND_V_MSG(get_child_count() == 0, FAILURE, "BT decorator has no child.");
 	int status = get_child(0)->execute(p_delta);
 	if (status == RUNNING) {
