+++ bt/decorators/bt_run_limit.cpp
@@ -6,7 +6,7 @@ String BTRunLimit::_generate_name() const {
 	return vformat("RunLimit x%d", run_limit);
 }
 
-int BTRunLimit::_tick(float p_delta) {
+int BTRunLimit::_tick(double p_delta) {
 	ERR_FAIL_COND_V_MSG(get_child_count() == 0, FAILURE, "BT decorator has no child.");
 	if (get_child(0)->get_status() != RUNNING) {
 		if (num_runs >= run_limit) {
