+++ bt/decorators/bt_time_limit.cpp
@@ -11,7 +11,7 @@ void BTTimeLimit::_enter() {
 	time_passed = 0.0;
 }
 
-int BTTimeLimit::_tick(float p_delta) {
+int BTTimeLimit::_tick(double p_delta) {
 	ERR_FAIL_COND_V_MSG(get_child_count() == 0, FAILURE, "BT decorator has no child.");
 	time_passed += p_delta;
 	int status = get_child(0)->execute(p_delta);
